[api] New, allow excluding columns from serialization
Marshmallow is incapable of operating on certain data types.
There is a mechanism for excluding columns from a model-based
API, but it does not work consistently when relationships are
defined and exposed by the API.

This change adds an additional mechanism for controlling which
columns are serialized using marshmallow by way of an attribute
on the model itself.

This is a more dependable and accessible mechanism, and the only
one that was found to work with relationships.

Closes #1126