[fix] redirect to next url after login
Currently, after login the user gets punted to the homepage / index.
Regardless of where they intended to go prior to having to login.
This change makes it so that the next parameter is actually used to
redirect users who have just logged in.
The is_safe_url() function aims to prevent phishing attempts, e.g.:
https://myinternalapp.example.co/login/?next=<externalurl>.