Fix, adapt marshmallow exclusion to v3
In newer marshmallow versions, passing a column name in the exclude
attribute without having it defined in the columns attribute will lead
to an error.
Previously, marshmallow would just ignore those cases.

If a field is manually excluded, then we need to remove that field name
from the list of exclusions for marshmallow to take care of.

See https://github.com/marshmallow-code/marshmallow/issues/877
for more information.